corrected the effects problem
