+++ js/create-preview.js
@@ -1,12 +1,10 @@
-// import {generateUsersPhotos} from './data.js';
 const templateFragment = document.querySelector('#picture').content;
 const template = templateFragment.querySelector('.picture');
 const listOfPictures = document.querySelector('.pictures');
 
+// Это функция конструктор? Тут заглавная буква?
 const renderListOfPictures = (usersPhotos) => {
   const picturesFragment = document.createDocumentFragment();
-
-  // usersPhotos = generateUsersPhotos(25);
 
   usersPhotos.forEach (({url,likes, comments}) => {
     const imageElement = template.cloneNode(true);
